Add a check against last page to show pagination component
There is an issue where if there is only one selectable page, the pagination component is hidden.  As a result if there are say...11 pages total, and the user is shown a max of 10 selectable pages at time, when they get to page 11, the component is only showing 1 selectable page, and the component hides.

Signed-off-by: seajamied <jdegnan@chef.io>